requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.16.2
anyio==4.9.0
apache-airflow==2.9.3
apache-airflow-providers-common-compat==1.6.1
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.26.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.3
apache-airflow-providers-http==5.2.2
apache-airflow-providers-imap==3.8.3
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-sqlite==4.0.2
apispec==6.8.2
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
boto3==1.38.42
boto3-stubs==1.38.45
botocore==1.38.42
botocore-stubs==1.38.30
cachelib==0.13.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
configupdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.4
deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
email-validator==2.2.0
flask==2.2.5
flask-appbuilder==4.5.3
flask-babel==2.0.0
flask-caching==2.3.1
flask-jwt-extended==4.7.1
flask-limiter==3.12
flask-login==0.6.3
flask-session==0.5.0
flask-sqlalchemy==2.5.1
flask-wtf==1.2.2
frozenlist==1.7.0
fsspec==2025.5.1
google-re2==1.1.20240702
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
inflection==0.5.1
itsdangerous==2.2.0
jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
limits==5.4.0
linkify-it-py==2.0.3
lockfile==0.12.2
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
multidict==6.5.0
mypy-boto3-s3==1.38.44
numpy==2.3.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
ordered-set==4.1.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.6.0
prison==0.2.1
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
py4j==0.10.9.7
pyarrow==20.0.0
pycparser==2.22
pygments==2.19.1
pyjwt==2.10.1
pyspark==3.5.6
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-nvd3==0.16.0
python-slugify==8.0.4
python-telegram-bot==22.1
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.1
rpds-py==0.25.1
ruff==0.12.0
s3transfer==0.13.0
setproctitle==1.3.6
six==1.17.0
sniffio==1.3.1
sqlalchemy==1.4.54
sqlalchemy-jsonfield==1.0.2
sqlalchemy-utils==0.41.2
sqlparse==0.5.3
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
types-awscrt==0.27.2
types-s3transfer==0.13.0
typing-extensions==4.14.0
tzdata==2025.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal-pathlib==0.2.6
urllib3==2.5.0
werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
wtforms==3.2.1
yarl==1.20.1
zipp==3.23.0
